refactor: use htt package for fetch requests
